chore: rename "kube-scanner" to "scanner"
this is a Kubernetes-centric repository, "kube" is redundant here
