+++ src/scanner/image/index.ts
