+++ src/scanner/index.ts
