+++ src/scanner/metadata-extractor.ts
