+++ src/scanner/types.ts
