+++ src/scanner/watchers/handlers/daemon-set.ts
