+++ src/scanner/watchers/handlers/types.ts
