+++ src/scanner/watchers/types.ts
