+++ src/scanner/workload-reader.ts
